// CODING/freelance/plut-mobile/src/shared/models/transactionModel.ts
// Generated by https://quicktype.io

import { BankAccountModel } from "./bankaccountModel";

export interface SellGiftCardModel {
  userId: string;
  bankAccountId: string;
  giftCardId: string;
  declaredAmount: number;
  countryCode: string;
  giftCardType: string;
  giftCardUrl: string;
  rateId: string;
}

export interface TransactionModel {
  cryptoExchanges: {
    declaredCryptoAmount: number;
    amountSentToCustomer: number;
    cryptoCurrencies: {
      logo: string;
      name: string;
      symbol: string;
    };
    exchangeRate: {
      quote: string;
      base: string;
      rate: number;
    };
    transactionHash: string;
  };
  id: "";
  failedReason: string;
  createdAt: string | number | Date;
  bankAccount: BankAccountModel;
  bankAccountId: string;
  transactionCategory: string;
  amount: number;
  currencyCode: string;
  userId: string;
  transactionStatus: string;
  giftCardExchanges: GiftCardExchange[];
}

export interface GiftCardExchange {
  transactionId: string;
  baseCurrencyCode: string;
  quoteCurrencyCode: string;
  amountInCard: number;
  declaredAmount: number;
  amountRecieved: number;
  giftCardUrl: string;
  giftCardId: string;
  giftCardRate: {
    rate: number;
    giftCardType: string;
    description?: string;
  };
  id: string;
  createdAt: Date;
  giftCard: {
    id: string;
    createdAt: string;
    name: string;
    imageUrl: string;
  };
  countryCode?: string;
}